# feat: metric selection for faster calculation