--- a/notebooks/example_abed.ipynb
+++ b/notebooks/example_abed.ipynb
@@ -10,7 +10,7 @@
     "import holoviews as hv\n",
     "import hvplot.pandas\n",
     "\n",
-    "from seastats.stats import get_stats"
+    "from seastats import get_stats"
    ]
   },
   {
@@ -54,8 +54,21 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "## Basic metrics\n",
-    "comparing `obs` pd.Series with `sim` pd.Series"
+    "## Suggested metrics (Default)\n",
+    "The metrics calculated by default is the addition of 2 lists:\n",
+    " * `GENERAL_METRICS`\n",
+    " * `STORM_METRICS`"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "from seastats import GENERAL_METRICS, STORM_METRICS\n",
+    "GENERAL_METRICS\n",
+    "STORM_METRICS"
    ]
   },
   {
@@ -72,6 +85,40 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
+    "we can get more metrics by using other lists instead:"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "from seastats import GENERAL_METRICS_ALL, STORM_METRICS_ALL\n",
+    "GENERAL_METRICS_ALL\n",
+    "STORM_METRICS_ALL"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "specifying `[\"all\"]` will retrieve all variables"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "pd.DataFrame(get_stats(sim, obs, metrics=[\"all\"]), index=[\"abed\"])"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
     "## Storm analysis"
    ]
   },
@@ -81,7 +128,6 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "from seastats.storms import storm_metrics\n",
     "from pyextremes import get_extremes\n",
     "\n",
     "quantile = 0.99\n",
@@ -197,7 +243,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "metrics = storm_metrics(sim, obs, quantile=0.99, cluster=72)\n",
+    "metrics = get_stats(sim, obs, quantile=0.99, cluster=72)\n",
     "pd.DataFrame(dict(stats, **metrics), index=['abed'])"
    ]
   }