Added new features: 
 * keyword `metrics` which can be `str` or `list`. Default `None` selects all metrics. 
 * added more thorough checks in tests
 * added duration print in pytest which gives the time spent for each function: 
 ```
 
=============================================================================================== test session starts ================================================================================================
platform linux -- Python 3.12.4, pytest-8.2.2, pluggy-1.5.0 -- /home/tomsail/work/python/seareport_org/seastats/.venv/bin/python
cachedir: .pytest_cache
rootdir: /home/tomsail/work/python/seareport_org/seastats
configfile: pyproject.toml
plugins: anyio-4.4.0
collected 5 items                                                                                                                                                                                                  

tests/compute_stats_test.py::test_get_all_stats_and_extremes PASSED                                                                                                                                          [ 20%]
tests/compute_stats_test.py::test_get_all_stats PASSED                                                                                                                                                       [ 40%]
tests/compute_stats_test.py::test_get_rmse PASSED                                                                                                                                                            [ 60%]
tests/compute_stats_test.py::test_get_slope_pp PASSED                                                                                                                                                        [ 80%]
tests/compute_stats_test.py::test_get_storm_metrics PASSED                                                                                                                                                   [100%]

================================================================================================ slowest durations =================================================================================================
0.26s call     tests/compute_stats_test.py::test_get_all_stats_and_extremes
0.24s call     tests/compute_stats_test.py::test_get_all_stats
0.07s call     tests/compute_stats_test.py::test_get_slope_pp
0.01s call     tests/compute_stats_test.py::test_get_storm_metrics
0.01s call     tests/compute_stats_test.py::test_get_rmse
0.00s setup    tests/compute_stats_test.py::test_get_all_stats_and_extremes
0.00s teardown tests/compute_stats_test.py::test_get_all_stats_and_extremes
0.00s teardown tests/compute_stats_test.py::test_get_all_stats
0.00s teardown tests/compute_stats_test.py::test_get_storm_metrics
0.00s teardown tests/compute_stats_test.py::test_get_rmse
0.00s setup    tests/compute_stats_test.py::test_get_all_stats
0.00s teardown tests/compute_stats_test.py::test_get_slope_pp
0.00s setup    tests/compute_stats_test.py::test_get_storm_metrics
0.00s setup    tests/compute_stats_test.py::test_get_rmse
0.00s setup    tests/compute_stats_test.py::test_get_slope_pp
```

NB: Calculating percentiles is the most intensive calculation, even more than POT selection.