@tomsail 2024-11-09
I have addressed the various things we've discussed yesterday. 
* made visible the `quantile` and `round` parameters. Both have default values (0 and -1) and deactivated.
* cleaned the tests and used `pytest.approx`
* using now by default `GENERAL_METRICS` and `STORM_METRIC` parameters lists for `get_stats()` and `storm_metrics()`. 
* vectorised the percentile calculation so it's now faster and made the `slope`/`intercept` API simpler